requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.10.1
setuptools==75.6.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
wheel==0.45.1
whitenoise==6.8.2
